requirements.txt
pydantic>=2.6
pydantic-settings>=2.5
python-dotenv>=1.0
pyyaml>=6.0
numpy>=2.2
pillow>=11.0
